Feat (game & UI): Added game page and basic logic
Added basic game page to test out reactive elements in Vue better and created a model to be able to communicate game logic between server and client.

Worked on a bunch of game logic and models to be able to represent a game of Yahtzee.

Will work on game model and game manager and build a prototype game to test client/server communication