Refactor (server): Lobby functionality
Changed the way lobbies, room, and now "channels" are handled in code. Mainly creating separate functions for each allowing you to join a room without necessarily it being a lobby.

A lobby is now the space that users can join and users will be in the lobby during the lifetime of the game they play.

A room just refers to rooms from Socket.IO (so the room(s) that a socket is in)

A channel is a socket.io room meant specifically for communication between clients — like a chat.

Lobbies are now also created by a user instead of using test lobbies.